HMRC-1438: Decouple FAQ Feedback model from Green Lanes
