+++ lib/trade_tariff_frontend.rb
@@ -114,6 +114,10 @@ module TradeTariffFrontend
     "Bearer #{ENV['GREEN_LANES_API_TOKEN']}"
   end
 
+  def backend_api_token
+    "Bearer #{ENV['BACKEND_API_TOKEN']}"
+  end
+
   def google_tag_manager_container_id
     ENV.fetch('GOOGLE_TAG_MANAGER_CONTAINER_ID', '')
   end
