+++ spec/models/green_lanes/faq_feedback_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe GreenLanes::FaqFeedback, type: :model do
+RSpec.describe FaqFeedback, type: :model do
   let(:faq_feedback) { build(:green_lanes_faq_feedback) }
 
   describe '#send_feedback_to_backend' do
